move deprecated to status module; deprecate old set_seed
